+++ src/main/kotlin/works/szabope/plugins/mypy/annotator/MypyAnnotator.kt
@@ -2,7 +2,7 @@ package works.szabope.plugins.mypy.annotator
 
 import com.intellij.openapi.project.Project
 import works.szabope.plugins.common.annotator.ToolAnnotator
-import works.szabope.plugins.common.services.ImmutableSettingsData
+import works.szabope.plugins.common.services.ToolExecutorConfiguration
 import works.szabope.plugins.mypy.services.MypyConfigurationResolver
 import works.szabope.plugins.mypy.services.MypySettings
 import works.szabope.plugins.mypy.services.SyncScanService
@@ -11,9 +11,9 @@ import works.szabope.plugins.mypy.services.parser.MypyMessage
 class MypyAnnotator : ToolAnnotator<MypyMessage>() {
     override val inspectionId = MypyInspectionId
 
-    override fun getSettingsInstance(project: Project) = MypySettings.getInstance(project)
+    override fun getSettings(project: Project) = MypySettings.getInstance(project)
 
-    override fun scan(info: AnnotatorInfo, configuration: ImmutableSettingsData): List<MypyMessage> {
+    override fun scan(info: AnnotatorInfo, configuration: ToolExecutorConfiguration): List<MypyMessage> {
         val resolvedConfig = MypyConfigurationResolver(info.project)
             .resolveForFile(info.file)
             .getOrElse { configuration }