+++ src/main/kotlin/works/szabope/plugins/mypy/dialog/DialogManager.kt
@@ -1,42 +1,42 @@
 // inspired by idea/243.19420.21 git4idea.DialogManager
 package works.szabope.plugins.mypy.dialog
 
-import com.intellij.openapi.ui.DialogWrapper
+import works.szabope.plugins.common.dialog.AbstractDialogManager
 import works.szabope.plugins.common.dialog.IDialogManager
-import works.szabope.plugins.common.dialog.IDialogManager.IShowDialog
 import works.szabope.plugins.common.dialog.PluginDialog
-import works.szabope.plugins.common.services.ImmutableSettingsData
 import works.szabope.plugins.common.services.PluginPackageManagementException
 
-private fun DialogWrapper.toMypyDialog() = object : PluginDialog {
-    override fun show() = this@toMypyDialog.show()
-}
-
-class DialogManager : IDialogManager {
-    override fun showDialog(dialog: PluginDialog) = dialog.show()
+class DialogManager : AbstractDialogManager() {
 
     override fun createPyPackageInstallationErrorDialog(exception: PluginPackageManagementException.InstallationFailedException) =
-        MypyPackageInstallationErrorDialog(exception.message).toMypyDialog()
+        MypyPackageInstallationErrorDialog(exception.message).asPluginDialog()
 
     override fun createToolExecutionErrorDialog(
-        configuration: ImmutableSettingsData,
+        commandLine: String,
         result: String,
-        resultCode: Int
-    ) = MypyExecutionErrorDialog(configuration, result, resultCode).toMypyDialog()
-
-    override fun createFailedToExecuteErrorDialog(message: String) =
-        FailedToExecuteErrorDialog(message).toMypyDialog()
+        resultCode: Int?
+    ) = MypyExecutionErrorDialog(commandLine, result, resultCode).asPluginDialog()
 
     override fun createToolOutputParseErrorDialog(
-        configuration: ImmutableSettingsData,
+        commandLine: String,
         targets: String,
         json: String,
         error: String
-    ) = MypyParseErrorDialog(configuration, targets, json, error).toMypyDialog()
+    ) = MypyParseErrorDialog(commandLine, targets, json, error).asPluginDialog()
 
-    override fun createGeneralErrorDialog(failure: Throwable) = MypyGeneralErrorDialog(failure).toMypyDialog()
+    override fun createGeneralErrorDialog(failure: Throwable) = MypyGeneralErrorDialog(failure).asPluginDialog()
 
-    companion object : IShowDialog {
-        override val dialogManager: IDialogManager by lazy { DialogManager() }
+    companion object : IDialogManager {
+        var dialogManager: IDialogManager = DialogManager()
+
+        override fun showDialog(dialog: PluginDialog) = dialogManager.showDialog(dialog)
+        override fun createPyPackageInstallationErrorDialog(exception: PluginPackageManagementException.InstallationFailedException) =
+            dialogManager.createPyPackageInstallationErrorDialog(exception)
+        override fun createToolExecutionErrorDialog(commandLine: String, result: String, resultCode: Int?) =
+            dialogManager.createToolExecutionErrorDialog(commandLine, result, resultCode)
+        override fun createToolOutputParseErrorDialog(commandLine: String, targets: String, json: String, error: String) =
+            dialogManager.createToolOutputParseErrorDialog(commandLine, targets, json, error)
+        override fun createGeneralErrorDialog(failure: Throwable) =
+            dialogManager.createGeneralErrorDialog(failure)
     }
 }