+++ src/main/kotlin/works/szabope/plugins/mypy/services/handleScanException.kt
@@ -3,27 +3,24 @@ package works.szabope.plugins.mypy.services
 import com.intellij.openapi.project.Project
 import kotlinx.coroutines.flow.FlowCollector
 import works.szabope.plugins.common.run.ToolExecutionTerminatedException
-import works.szabope.plugins.common.services.ImmutableSettingsData
+import works.szabope.plugins.common.services.IncompleteConfigurationNotifier
+import works.szabope.plugins.common.services.showClickableBalloonError
 import works.szabope.plugins.mypy.MypyBundle
 import works.szabope.plugins.mypy.dialog.DialogManager
+import works.szabope.plugins.mypy.toolWindow.MypyToolWindowPanel
 
 inline fun <reified T> handleScanException(
-    project: Project, configuration: ImmutableSettingsData, stdErr: StringBuilder
+    project: Project, noinline commandLine: () -> String?, stdErr: StringBuilder,
+    notifier: IncompleteConfigurationNotifier, silent: Boolean = false
 ): suspend FlowCollector<T>.(Throwable) -> Unit = {
     if (it is ToolExecutionTerminatedException) {
-        showClickableBalloonError(project, MypyBundle.message("mypy.toolwindow.balloon.external_error")) {
+        if (!silent) showClickableBalloonError(project, MypyToolWindowPanel.ID, MypyBundle.message("mypy.toolwindow.balloon.external_error")) {
             DialogManager.showToolExecutionErrorDialog(
-                configuration, stdErr.toString(), it.exitCode
+                commandLine() ?: "", stdErr.toString(), it.exitCode
             )
         }
     } else {
-        // Unexpected exception
-        showClickableBalloonError(
-            project, MypyBundle.message("mypy.toolwindow.balloon.failed_to_execute")
-        ) {
-            DialogManager.showFailedToExecuteErrorDialog(
-                it.message ?: MypyBundle.message("mypy.please_report_this_issue")
-            )
-        }
+        // Unexpected exception - tool likely gone
+        notifier.showWarningBubble(false)
     }
 }