+++ src/main/kotlin/works/szabope/plugins/mypy/services/parser/MypyMessage.kt
@@ -8,10 +8,10 @@ import works.szabope.plugins.common.annotator.ToolMessage
 @Serializable
 data class MypyMessage(
     val file: String,
-    override var line: Int,
+    override val line: Int,
     override val column: Int,
     override val message: String,
-    val hint: String?,
+    val hint: String? = null, // without initialization, it is still treated as required; alternatively we could set explicitNulls on parser
     val code: String,
-    var severity: String
+    val severity: String
 ) : ToolMessage