+++ src/test/kotlin/works/szabope/plugins/mypy/action/RescanTest.kt
@@ -6,6 +6,7 @@ import com.intellij.testFramework.TestDataPath
 import works.szabope.plugins.mypy.AbstractToolWindowTestCase
 import works.szabope.plugins.mypy.services.MypySettings
 import works.szabope.plugins.mypy.testutil.dataContext
+import works.szabope.plugins.common.test.action.invokeNamedActionWithScope
 import works.szabope.plugins.mypy.testutil.scan
 import java.nio.file.Paths
 import kotlin.io.path.absolutePathString
@@ -24,7 +25,7 @@ class RescanTest : AbstractToolWindowTestCase() {
         }
         val file = myFixture.configureByText("a.py", "doesn't matter").virtualFile
         scan(dataContext(project) { add(CommonDataKeys.VIRTUAL_FILE_ARRAY, arrayOf(file)) })
-        PlatformTestUtil.waitWhileBusy { ScanJobRegistry.INSTANCE.isActive() }
+        PlatformTestUtil.waitWhileBusy { MypyScanJobRegistryService.getInstance(project).isActive() }
     }
 
     /**
@@ -33,8 +34,8 @@ class RescanTest : AbstractToolWindowTestCase() {
      */
     fun `test rescan running for the same file scan did`() {
         MypySettings.getInstance(project).executablePath = Paths.get(testDataPath).resolve("mypy2").absolutePathString()
-        PlatformTestUtil.invokeNamedAction(RescanAction.ID)
-        PlatformTestUtil.waitWhileBusy { ScanJobRegistry.INSTANCE.isActive() }
+        invokeNamedActionWithScope(RescanAction.ID)
+        PlatformTestUtil.waitWhileBusy { MypyScanJobRegistryService.getInstance(project).isActive() }
         treeUtil.assertStructure("+Found 1 issue(s) in 1 file(s)\n")
         treeUtil.expandAll()
         treeUtil.assertStructure(